split case importer task status models out of jsmodels.py
